BL-527 : Fixing nested tag removal in active tag. Issue reported as #98 in Reznok branch (https://github.com/reznok/GMCAbilitySystem/issues/98)
--------------------------------------------

Issue was GetActiveEffectByTag() was looking also for depth in tag hierarchy. Fixed by adding an optional argument (bMatchExact), this argument has been set by default to true, and can mess with project who are calling this function if they are excepting in depht.

Not ideal for retrocompability, but i prefere to enforce a good practice over the time.